+++ tests/scripts/script4.py
@@ -3,7 +3,10 @@ import sys
 env = env  # noqa
 
 
-op = sys.argv[1]
+op = ''
+if len(sys.argv) > 1:
+    op = sys.argv[1]
+
 env['ir.config_parameter'].set_param('testparam', 'testvalue')
 if op == 'commit':
     env.cr.commit()